moving function to an hook
